api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.10
pydantic==1.10.13
python-multipart==0.0.6
pytz==2023.3
